Factor out function looking for devcontainer files to devcontainer package
